Add stringified backtrace to WrappedError passed back to Lua
This makes the behaviour consistent with how Lua behaves when a
normal string is used as error instead of a userdata.

When a string is used, the Lua runtime will add a complete
stacktrace to the error output; when a userdata is used, it will
not do that, but only display the tostring(ud) at the time the
error is printed.

This makes debugging things much harder (especially when writing
modules), because errors emitted by the Rust code (such as type
conversion errors) are not attribute to a code line.

However, this implementation has two important caveats:

- The runtime complexity of walking the stack and transforming it
  into a nice human-readable string; this complexity is to be
  expected to also exist in case of Lua doing that internally,
  modulo extra cost for utf8 checks inside Rust.

- When the WrappedError is transported back into and out of Rust,
  the stacktrace is lost and regenerated when passed back into
  Lua.

  This means that the error cause can be "shadowed" by passing
  through Rust.

  Handling this would require a more invasive change which allows
  chaining Error instances in a way which allows carrying the
  backtrace without changing the Error enum to contain an optional
  stacktrace in each variant; I’ll leave this to future work.

Fixes #44.