userdata: Add support for IPairs metamethod
This is available in Lua 5.2 [1] and Lua 5.3, but not in
Lua 5.4 or Lua 5.1.

Lua 5.3 deprecated `__ipairs` but still uses it when available
(seems to even require it for userdata objects based on my tests)
while Lua 5.4 replaced the `ipairs()` behaviour with counting
upwards from 1 and indexing the object (respecting `__index`)
until a nil value is found.

   [1]: https://www.lua.org/manual/5.2/manual.html#pdf-ipairs