WIP: Add Lua::with_debug_iter for arbitrary access of stack info
This adds `Lua::with_debug_iter` which takes a closure that takes a lending
iterator of `Debug` instances as an argument. This is an unfortunate artifact
of the current API design, with `Debug` borrowing, instead of owning the
underlying 'lua_Debug` C struct. It also sets a bogus `LUA_HOOKCALL` value
to the `event` field, since the current signature of `Debug::event` only
considered the case of hook argument.

The intended use case is exposing custom log functions to Lua code.

A more natural API could potentially be exposed if some breaking changes
could be made to the `Debug` type to make it own or be generic over the
`lua_Debug` struct.